Move creation of sub-company to the template-editor.cases.js [stage-3]
